[Admin][CatalogPromotion] Remove GenerateFormElement
