+++ src/Sylius/Behat/Page/Admin/PromotionCoupon/GeneratePageInterface.php
@@ -13,9 +13,23 @@ declare(strict_types=1);
 
 namespace Sylius\Behat\Page\Admin\PromotionCoupon;
 
-use FriendsOfBehat\PageObjectExtension\Page\SymfonyPageInterface;
+use Sylius\Behat\Page\Admin\Crud\CreatePageInterface as BasePageInterface;
 
-interface GeneratePageInterface extends SymfonyPageInterface
+interface GeneratePageInterface extends BasePageInterface
 {
     public function generate(): void;
+
+    public function specifyPrefix(string $prefix): void;
+
+    public function specifyCodeLength(?int $codeLength): void;
+
+    public function specifySuffix(string $suffix): void;
+
+    public function specifyAmount(?int $amount): void;
+
+    public function setExpiresAt(\DateTimeInterface $date): void;
+
+    public function setUsageLimit(int $limit): void;
+
+    public function getFormValidationMessage(): string;
 }
