+++ src/Sylius/Behat/Resources/config/services/contexts/ui.xml
@@ -270,7 +270,6 @@
             <argument type="service" id="sylius.behat.page.admin.promotion_coupon.index" />
             <argument type="service" id="sylius.behat.page.admin.promotion_coupon.update" />
             <argument type="service" id="sylius.behat.element.admin.promotion_coupon.form" />
-            <argument type="service" id="sylius.behat.element.admin.promotion_coupon.generate_form" />
             <argument type="service" id="sylius.behat.notification_checker.admin" />
         </service>
