+++ src/Sylius/Behat/Resources/config/services/elements/admin.xml
@@ -93,11 +93,5 @@
             class="Sylius\Behat\Element\Admin\PromotionCoupon\FormElement"
             parent="sylius.behat.element.admin.crud.form"
         />
-
-        <service
-            id="sylius.behat.element.admin.promotion_coupon.generate_form"
-            class="Sylius\Behat\Element\Admin\PromotionCoupon\GenerateFormElement"
-            parent="sylius.behat.element.admin.crud.form"
-        />
     </services>
 </container>
