+++ src/Sylius/Behat/Resources/config/services/pages/admin/promotion_coupon.xml
@@ -25,7 +25,9 @@
         <service id="sylius.behat.page.admin.promotion_coupon.create" class="%sylius.behat.page.admin.promotion_coupon.create.class%" parent="sylius.behat.page.admin.crud.create" public="false">
             <argument type="string">sylius_admin_promotion_coupon_create</argument>
         </service>
-        <service id="sylius.behat.page.admin.promotion_coupon.generate" class="%sylius.behat.page.admin.promotion_coupon.generate.class%" parent="sylius.behat.symfony_page" public="false" />
+        <service id="sylius.behat.page.admin.promotion_coupon.generate" class="%sylius.behat.page.admin.promotion_coupon.generate.class%" parent="sylius.behat.page.admin.crud.create" public="false">
+            <argument type="string">sylius_admin_promotion_coupon_generate</argument>
+        </service>
         <service id="sylius.behat.page.admin.promotion_coupon.index" class="%sylius.behat.page.admin.promotion_coupon.index.class%" parent="sylius.behat.page.admin.crud.index" public="false">
             <argument type="string">sylius_admin_promotion_coupon_index</argument>
         </service>
